Refactor OAuth browser login flow
Refactor OAuth browser login flow so that it is available in both interactive (UI) and batch (CLI) modes.
